Fix crash in serialization of data in tags cache on 64b machines.
git-svn-id: svn://svn.daper.net/moc/trunk@2097 910807d9-36e0-0310-a014-e9ea483e2ba4

+++ tags_cache.c
@@ -165,7 +165,7 @@ static char *cache_record_serialize (const struct cache_record *rec, int *len)
 
 	*len = sizeof(rec->mod_time)
 		+ sizeof(rec->atime)
-		+ sizeof(int) * 3 /* lenghts of title, artist, time. */
+		+ sizeof(size_t) * 3 /* lenghts of title, artist, time. */
 		+ artist_len
 		+ album_len
 		+ title_len
